Fixed dropped packets on NRF52
This was a very pesky little bug. The previous implementation relied upon the state of the IN register, which was unreliable if the gpio irq was late i.e. if the line was low for a short amount of time, the line might be high at the time the interrupt is handled resulting in a missed edge event.

This commit uses the LATCH register for determining if a transition has occurred, this is infinitely more reliable and I should read the datasheet better next time...