Add PPI peripheral and use it in SWS
For some reason the nordic engineers decided to not make it possible to retrieve the number of bytes transferred in an easy dma transaction.

In this commit, I added a PPI linked from TXD / RXD to a timer count event to count the number of bytes transferred. This is not strictly necessary, but it's certainly cool.