Explicitly set SAADC channel config mode to single ended.
The SE (single ended) mode is set with value 0 at bit 20 in the
CONFIG register for each channel. So, by default when this bit is
not set (as it was before this commit) it will be configured to SE.

This commit does not change the value set to `CH[n].CONFIG`, but
it's better to be explicit about the configuration.